Spring boot security jwt + some clients + transfer

+++ src/main/java/kz/alimgaziyev/bankappspringbooth2/dto/AccountDto.java
@@ -8,7 +8,7 @@ import lombok.NoArgsConstructor;
 @AllArgsConstructor
 @Data
 public class AccountDto {
-    Integer clientId;
+    Long clientId;
     String accountType;
 
     public String getClientId() {
